cli: move `RepoBranchName(Pattern)` to `crate::cli_utils`
Needed in another upcoming diff, as Gerrit needs to parse remote names for
pushing refs.
